Feat: 할일뷰 최종디자인에 따른 구조 수정
