+++ OneByte/Source/Features/Work/View/TodayRoutineView.swift
@@ -1,0 +1,21 @@
+//
+//  TodayRoutineView.swift
+//  OneByte
+//
+//  Created by 이상도 on 11/19/24.
+//
+
+import SwiftUI
+
+struct TodayRoutineView: View {
+    var body: some View {
+        VStack {
+            Text("Hello, World!")
+        }
+        .frame(maxWidth: .infinity)
+    }
+}
+
+#Preview {
+    TodayRoutineView()
+}
