+++ OneByte/Source/Features/Work/View/WorkView.swift
@@ -8,90 +8,90 @@
 import SwiftUI
 import SwiftData
 
-enum tapInfo : String, CaseIterable {
-    
-    case first, second, third, fourth
-    
-    var colorClovar: String {
-        switch self {
-        case .first:
-            return "ColorClover1"
-        case .second:
-            return "ColorClover2"
-        case .third:
-            return "ColorClover3"
-        case .fourth:
-            return "ColorClover4"
-        }
-    }
-    
-    var grayClovar: String {
-        switch self {
-        case .first:
-            return "GrayClover1"
-        case .second:
-            return "GrayClover2"
-        case .third:
-            return "GrayClover3"
-        case .fourth:
-            return "GrayClover4"
-        }
-    }
+enum routineTapInfo : String, CaseIterable {
+    case today = "오늘의 루틴"
+    case all = "전체 루틴"
 }
 
+// MARK: 루틴 메인 뷰
 struct WorkView: View {
     
-    @State private var selectedPicker: tapInfo = .first
+    @State private var selectedPicker: routineTapInfo = .today
     @Namespace private var animation
-    @Query var mainGoals: [MainGoal]
+    var routineType : routineTapInfo
     
     var body: some View {
-        VStack(spacing: 0) {
-            HStack {
-                Text("할 일")
-                    .font(.Pretendard.Bold.size22)
-                    .foregroundStyle(Color(hex: "B4A99D"))
-                
-                Spacer()
-                
-                NavigationLink {
-                    SettingView()
-                } label: {
-                    Image(systemName: "info.circle")
-                        .resizable()
-                        .frame(width: 24, height: 24)
-                        .foregroundStyle(Color(hex: "B4A99D"))
+        ZStack {
+            Color.my6FB56F
+                .ignoresSafeArea(edges: .top)
+            
+            VStack(spacing: 0) {
+                HStack {
+                    Text("할 일")
+                        .font(.Pretendard.Bold.size22)
+                        .foregroundStyle(.white)
+                    
+                    Spacer()
+                    
+                    NavigationLink {
+                        SettingView()
+                    } label: {
+                        Image(systemName: "gear")
+                            .resizable()
+                            .frame(width: 24, height: 24)
+                            .foregroundStyle(Color.myCEEDCE)
+                    }
                 }
+                .padding()
+                
+                HStack{
+                    Text("완벽하지 않아도 괜찮아요.\n중요한 것은 꾸준히 다시 시작하는 거예요!")
+                    Image("Turtle_Main")
+                        .resizable()
+                        .scaledToFit()
+                        .frame(width: 105, height: 85)
+                }
+                .padding()
+                .frame(maxWidth: .infinity)
+                .background(Color.my6FB56F)
+                
+                VStack(spacing: 0) {
+                    animate() // Tabbar Picker
+                    Divider()
+                        .foregroundStyle(Color.myF0E8DF)
+                    
+                    ScrollView(.vertical, showsIndicators: false) { // Picker에 따른 뷰
+                        switch selectedPicker {
+                        case .today:
+                            TodayRoutineView()
+                        case .all:
+                            AllRoutineView()
+                        }
+                    }
+                }
+                .background(Color.myFFFAF4)
+                .clipShape(RoundedCorner(radius: 12, corners: [.topLeft, .topRight]))
             }
-            .padding()
-            
-            SubgoalTabView()
-            
-            Divider()
-                .foregroundStyle(Color.myF0E8DF)
-            
-            if let mainGoal = mainGoals.first,
-               let selectedSubGoal = mainGoal.subGoals.first(where: { $0.id == subGoalId(for: selectedPicker) }) {
-                TodoView(tests: selectedPicker, subGoal: selectedSubGoal)
-            } else {
-                Text("현재 Subgoal 데이터가 없습니다.")
-            }
-            Spacer()
         }
-        .ignoresSafeArea(edges: .bottom)
     }
     
     @ViewBuilder
-    private func SubgoalTabView() -> some View {
+    private func animate() -> some View {
         HStack {
-            ForEach(tapInfo.allCases, id: \.self) { item in
-                VStack(spacing: 0) {
-                    Image(selectedPicker == item ? item.colorClovar : item.grayClovar)
-                        .resizable()
-                        .scaledToFit()
-                        .frame(maxWidth: .infinity/4)
-                        .frame(height: 51)
-                        .padding()
+            ForEach(routineTapInfo.allCases, id: \.self) { item in
+                VStack {
+                    Text(item.rawValue)
+                        .font(.Pretendard.Bold.size17)
+                        .frame(maxWidth: .infinity/4, minHeight: 50)
+                        .foregroundStyle(selectedPicker == item ? Color.my1D1D1D : .gray)
+                    
+                    if selectedPicker == item {
+                        Capsule()
+                            .foregroundStyle(Color.my95D895)
+                            .frame(height: 2)
+                            .padding(.horizontal, 40)
+                            .matchedGeometryEffect(id: "info", in: animation)
+                    }
                 }
                 .onTapGesture {
                     withAnimation(.easeInOut) {
@@ -101,164 +101,8 @@ struct WorkView: View {
             }
         }
     }
-    
-    // 탭 선택에 따른 SubGoal ID 반환
-    private func subGoalId(for tab: tapInfo) -> Int {
-        switch tab {
-        case .first: return 1
-        case .second: return 2
-        case .third: return 3
-        case .fourth: return 4
-        }
-    }
-}
-
-struct TodoView : View {
-    
-    var tests : tapInfo
-    var subGoal: SubGoal
-    
-    var body: some View {
-        VStack {
-            ScrollView(.vertical, showsIndicators: false) {
-                HStack {
-                    Image(tests.colorClovar)
-                        .resizable()
-                        .frame(width: 29, height: 29)
-                    
-                    Text(subGoal.title == "" ? "서브목표가 비어있어요." : subGoal.title)
-                        .font(.Pretendard.Bold.size22)
-                        .foregroundStyle(Color.my2B2B2B)
-                    Spacer()
-                }
-                .padding([.top, .horizontal])
-                
-                VStack(spacing: 16) {
-                    // 빈 제목이 아닌 DetailGoal만 표시
-                    ForEach(subGoal.detailGoals.filter { !$0.title.isEmpty }, id: \.id) { detailGoal in
-                        WeekAchieveCell(
-                            detailGoalTitle: detailGoal.title,
-                            achieveCount: detailGoal.achieveCount,
-                            achieveGoal: detailGoal.achieveGoal,
-                            alertMon: detailGoal.alertMon,
-                            alertTue: detailGoal.alertTue,
-                            alertWed: detailGoal.alertWed,
-                            alertThu: detailGoal.alertThu,
-                            alertFri: detailGoal.alertFri,
-                            alertSat: detailGoal.alertSat,
-                            alertSun: detailGoal.alertSun,
-                            isRemind: detailGoal.isRemind,
-                            remindTime: detailGoal.remindTime,
-                            achieveMon: detailGoal.achieveMon,
-                            achieveTue: detailGoal.achieveTue,
-                            achieveWed: detailGoal.achieveWed,
-                            achieveThu: detailGoal.achieveThu,
-                            achieveFri: detailGoal.achieveFri,
-                            achieveSat: detailGoal.achieveSat,
-                            achieveSun: detailGoal.achieveSat
-                        )
-                    }
-                }
-                .padding(.horizontal)
-            }
-        }
-        .frame(maxWidth: .infinity)
-        .background(Color.myFFFAF4)
-    }
-}
-
-struct WeekAchieveCell: View {
-    
-    let detailGoalTitle: String // DetailGoal 제목
-    let achieveCount: Int // 현재 달성 횟수
-    let achieveGoal: Int // 목표 달성 횟수
-    let alertMon: Bool
-    let alertTue: Bool
-    let alertWed: Bool
-    let alertThu: Bool
-    let alertFri: Bool
-    let alertSat: Bool
-    let alertSun: Bool
-    let isRemind: Bool
-    let remindTime: Date?
-    var achieveMon: Bool
-    var achieveTue: Bool
-    var achieveWed: Bool
-    var achieveThu: Bool
-    var achieveFri: Bool
-    var achieveSat: Bool
-    var achieveSun: Bool
-    
-    let days: [String] = ["월","화","수","목","금","토","일"] // 요일 데이터 - 고정
-    
-    // 실제 현재 요일 (월, 화, 수, ...)
-    private var currentDay: String {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "ko_KR")
-        formatter.dateFormat = "E" // 요일만 가져오기
-        return formatter.string(from: Date())
-    }
-    
-    var body: some View {
-        VStack(spacing: 15) {
-            VStack(spacing: 5) {
-                HStack {
-                    Text(detailGoalTitle)
-                        .font(.Pretendard.Bold.size16)
-                        .foregroundStyle(Color.my2B2B2B)
-                    
-                    Spacer()
-                    
-                    Text("달성한 횟수 \(achieveCount)/\(achieveGoal)")
-                        .font(.Pretendard.Medium.size12)
-                        .foregroundStyle(Color.my727272)
-                }
-                HStack {
-                    if let remindTime = remindTime {
-                        Text(remindTime.alertTimeString)
-                            .font(.Pretendard.Medium.size12)
-                            .foregroundStyle(Color.my727272)
-                    } else {
-                        Text("")
-                    }
-                    Spacer()
-                }
-            }
-            HStack {
-                ForEach(0..<days.count, id: \.self) { index in
-                    VStack(spacing: 4) {
-                        Text(days[index])
-                            .font(.Pretendard.Medium.size11)
-                            .frame(width: 18, height: 18)
-                            .foregroundStyle(currentDay == days[index] ? .white : Color.my7D7D7D)
-                            .background(currentDay == days[index] ? Color.my385E38 : .clear)
-                            .clipShape(Circle())
-                        
-                        Button {
-                            print("클로버 버튼 탭")
-                        } label: {
-                            Rectangle()
-                        }
-                        .frame(maxWidth: .infinity/7, minHeight: 41) // ⚠️⚠️ 나중에 높이 수정
-                        .foregroundColor(currentDay == days[index] ? .white : Color.myDBDBDC)
-                        .overlay(currentDay == days[index] ? RoundedRectangle(cornerRadius: 8).stroke(Color.my385E38, lineWidth: 3) : nil)
-                        .cornerRadius(8)
-                    }
-                    .frame(maxWidth: .infinity)
-                }
-            }
-            .frame(maxWidth: .infinity)
-        }
-        .padding()
-        .background(Color.white)
-        .cornerRadius(12)
-        .overlay(
-            RoundedRectangle(cornerRadius: 12)
-                .stroke(Color.myF0E8DF, lineWidth: 1)
-        )
-    }
 }
 
 #Preview {
-    WorkView()
+    WorkView(routineType: .today)
 }
